packages/nde-erfgoed-manage/lib/app.root.spec.ts
import { AppRootComponent } from './app.root';

describe('AppRootComponent', () => {
  let component: AppRootComponent;

  beforeEach(() => {
    component = window.document.createElement('nde-app-root') as AppRootComponent;
    (component as any).createRenderRoot = () => this;
  });

  afterEach(() => {
    document.getElementsByTagName('html')[0].innerHTML = '';
  });

  it('should be correctly instantiated', () => {
    expect(component).toBeTruthy();
  });

  it('should show alerts when set', async () => {
    component.alerts = [ {
      type: 'success',
      message: 'Foo',
    } ];

    window.document.body.appendChild(component);
    await component.updateComplete;

    const alerts = window.document.body.getElementsByTagName('nde-app-root')[0].shadowRoot.querySelectorAll('nde-alert');

    expect(alerts).toBeTruthy();
    expect(alerts.length).toBe(1);
  });

  it('should not show alerts when unset', async () => {
    component.alerts = null;

    window.document.body.appendChild(component);
    await component.updateComplete;

    const alerts = window.document.body.getElementsByTagName('nde-app-root')[0].shadowRoot.querySelectorAll('nde-alert');

    expect(alerts).toBeTruthy();
    expect(alerts.length).toBe(0);
  });
});
